LanguageSettings kan jo fint vera internal
